Rename: 데이터베이스 예약어 충돌로 인해 User -> Member 변경
