+++ src/main/java/com/alzzaipo/domain/member/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.alzzaipo.domain.user;
+package com.alzzaipo.domain.member;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
